Rename tests namespace
