+++ tests/Provider/HostName.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Kassner\Tests\LogParser\Provider;
+namespace Kassner\LogParser\Tests\Provider;
 
 class HostName extends \PHPUnit_Framework_TestCase
 {
